+++ amazonorders/entity/shipment.py
@@ -3,9 +3,9 @@ from typing import List, Optional
 
 from bs4 import Tag
 
+from amazonorders.constants import BASE_URL
 from amazonorders.entity.item import Item
 from amazonorders.entity.parsable import Parsable
-from amazonorders.session import BASE_URL
 
 __author__ = "Alex Laird"
 __copyright__ = "Copyright 2024, Alex Laird"